model id datatype same as db datatype

+++ backend/src/main/java/com/odde/doughnut/models/Note.java
@@ -10,8 +10,6 @@ import lombok.Setter;
 @Entity
 @Table(name = "note")
 public class Note {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+  @Id @GeneratedValue(strategy = GenerationType.IDENTITY) private Integer id;
   @Setter private String title;
 }
